Simplify logging infrastructure in `usb_fmt`
Makes the `write` method non-blocking, giving the caller more
flexibility regarding what should happen when the buffer is full. For
example, a caller might decide to drop the message, if blocking is not
acceptable.

Adds a new method, `write_blocking`, that behaves similarly to how
`write` behaved before, although in a less sophisticated manner.